+++ docs/changes.rst
@@ -5,6 +5,7 @@ Not yet released.
 
 **New features**
 
+* Added :ref:`component-key_filter` in the component.
 * :ref:`Searching` now supports filtering by object path and :ref:`date-search`.
 * Merge requests credentials can now be passed in the repository URL, see :ref:`settings-credentials`.
 * :ref:`mt-azure-openai` automatic suggestion service.